fix(#380): CodeQL整数オーバーフロー警告を修正
IoU計算のunionArea算出で整数乗算がfloatキャスト前にオーバーフロー
する可能性をCodeQLが検出。乗算前にfloatキャストすることで解消。
foreachの暗黙フィルタをWhere句に変換。

+++ Baketa.Infrastructure/Processing/Strategies/OcrExecutionStageStrategy.cs
@@ -1535,7 +1535,7 @@ public class OcrExecutionStageStrategy : IProcessingStageStrategy
             return 0f;
 
         var intersectionArea = (float)(intersectRight - intersectX) * (intersectBottom - intersectY);
-        var unionArea = (float)(a.Width * a.Height + b.Width * b.Height) - intersectionArea;
+        var unionArea = (float)a.Width * a.Height + (float)b.Width * b.Height - intersectionArea;
 
         return unionArea > 0 ? intersectionArea / unionArea : 0f;
     }
